requirements/base.txt
pyjwt[crypto]
itsdangerous
mjml
aiosmtplib
aiohttp==3.10.10
alembic
amqp
annotated-types
argon2-cffi
asgiref
asyncpg
babel==2.16.0
bcrypt==4.0.1
ccxt==4.4.23
celery
cloudinary
dnspython
email-validator
fastapi
fastapi-cli
fastapi-mail
fastapi-pagination==0.12.31
hiredis
humanize
jinja2
loguru
mjml-python
pandas-ta==0.3.14b
passlib
paystackapi==2.1.3
phonenumbers==8.13.47
pillow
pycountry==24.6.1
pydantic
pydantic-settings
pydantic-extra-types==2.9.0
python-dateutil
python-jose
python-magic
python-multipart
python-slugify
redis
resend
requests
sqlalchemy
sqlmodel
typing-extensions
uvicorn
uvicorn-worker
urllib3
websockets
yfinance==0.2.44
yahooquery==2.3.7
